Fix compatibility issue with 64-bit cygwin.

+++ src/support/os_cygwin.cpp
@@ -449,7 +449,7 @@ bool autoOpenFile(string const & filename, auto_open_mode const mode,
 	// reference: http://msdn.microsoft.com/en-us/library/bb762153.aspx
 	string const win_path = to_local8bit(from_utf8(convert_path(filename, PathStyle(windows))));
 	char const * action = (mode == VIEW) ? "open" : "edit";
-	bool success = reinterpret_cast<int>(ShellExecute(NULL, action,
+	bool success = reinterpret_cast<long>(ShellExecute(NULL, action,
 					win_path.c_str(), NULL, NULL, 1)) > 32;
 
 	if (!path.empty() && !lyxrc.texinputs_prefix.empty()) {
